переработаны аннотации - добавлена возможность передавать данные для создания, выбор добавлять в очистку или нет

+++ src/main/java/jupiter/extension/ProjectExtension.java
@@ -4,6 +4,7 @@ import api.models.CreateProjectRequest;
 import api.models.CreateUserResponse;
 import api.models.ParentProject;
 import api.requests.steps.ProjectManagementSteps;
+import common.data.ProjectData;
 import common.generators.TestDataGenerator;
 import jupiter.annotation.WithProject;
 import org.junit.jupiter.api.extension.*;
@@ -26,24 +27,32 @@ public class ProjectExtension implements BeforeEachCallback, ParameterResolver {
                 throw new ExtensionConfigurationException("Добавьте аннотацию @WithUsersQueue");
             }
 
+            String parentProjectId = annotation.parentProjectId().equals("default") ? ProjectData.PARENT_PROJECT.value : annotation.parentProjectId();
+            String projectId = annotation.projectId().equals("default") ? TestDataGenerator.generateProjectID() : annotation.projectId();
+            String projectName = annotation.projectName().equals("default") ? TestDataGenerator.generateProjectName() : annotation.projectName();
+
             CreateProjectRequest project = CreateProjectRequest.builder()
-                    .id(TestDataGenerator.generateProjectID())
-                    .name(TestDataGenerator.generateProjectName())
-                    .parentProject(new ParentProject(annotation.parentProjectId().value))
+                    .id(projectId)
+                    .name(projectName)
+                    .parentProject(new ParentProject(parentProjectId))
                     .build();
 
-            ProjectManagementSteps.createProject(
-                    project.getId(),
-                    project.getName(),
-                    project.getParentProject(),
-                    user
-            );
+            if(!annotation.useExisting()) {
+                ProjectManagementSteps.createProject(
+                        project.getId(),
+                        project.getName(),
+                        project.getParentProject(),
+                        user
+                );
+            }
 
             context.getStore(NAMESPACE).put(context.getUniqueId(), project);
 
-            context.getStore(NAMESPACE).put("project_cleanup", (ExtensionContext.Store.CloseableResource) () -> {
+            if(annotation.addToCleanup() && !annotation.useExisting()) {
+                context.getStore(NAMESPACE).put("project_cleanup", (ExtensionContext.Store.CloseableResource) () -> {
                     ProjectManagementSteps.deleteProjectByIdQuietly(project.getId(), user);
-            });
+                });
+            }
         }
     }
